Add ACK socket.io support:
socket.emitWithAck("message", ["Hello world!"]).then( (data) {
  // this callback runs when this specific message is acknowledged by the server
  print(data);
});

When ACK is requested, a Future is returned, which will be completed once the
server acknowledges the request. Note that there is no guarantee that the server
will ever acknowledge the request, and the current implementation does not provide
any timeout or garbage collection mechanism.